[BSK-1403] The last argument of linspace must be an integer.  With the "/" operator in python now making the result a float, I have cast the answer as an integer to avoid a pytest warning about depreciated functionality.
Former-commit-id: 0dca7466de412b1b809ea8d1925ed014f8d13d01
Former-commit-id: 35580b4f86070489c52233745491a53c035b24e3

+++ src/fswAlgorithms/opticalNavigation/pixelLineBiasUKF/_UnitTest/test_pixelLineBiasUKF.py
@@ -286,7 +286,7 @@ def StatePropRelOD(show_plots):
     unitTestSim.ConfigureStopTime(macros.min2nano(timeSim))
     unitTestSim.ExecuteSimulation()
 
-    time = np.linspace(0,timeSim*60,timeSim*60/dt+1)
+    time = np.linspace(0, timeSim*60, (int) (timeSim*60/dt+1))
     dydt = np.zeros(len(moduleConfig.stateInit))
     energy = np.zeros(len(time))
     expected=np.zeros([len(time), len(moduleConfig.stateInit)+1])
